requirements.txt
opencv_python_headless==4.5.2.52
torchvision==0.8.1
scipy==1.5.0
requests==2.24.0
pandas==1.0.5
numpy==1.18.5
torch==1.7.0
seaborn==0.10.1
tqdm==4.47.0
matplotlib==3.2.2
streamlit==0.73.1
pafy==0.5.5
Pillow==8.2.0
protobuf==3.17.3
PyYAML==5.4.1
thop==0.0.31-2005241907
